Alignment enforces validity of arguments
Had to learn how to raise exceptions and how to expect them in rspec.
Learned about the syntax for writing a setter too, kinda cool.
Also, is it a good approach to define the valid list in the method like
I'm doing? Should that list be a constant somewhere so it doesn't have
to be created each time the method is called? I'm sure it should, but
what's the Ruby Way of doing that?